refactor : fix lines and imports
