+++ src/main/java/codesquad/controller/ApiAnswerController.java
@@ -1,7 +1,7 @@
 package codesquad.controller;
 
 import codesquad.model.answer.AnswerDto;
-import codesquad.model.answer.ContentDto;
+import codesquad.model.content.Content;
 import codesquad.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,13 +16,13 @@ public class ApiAnswerController {
     private QuestionService questionService;
 
     @PostMapping("")
-    public AnswerDto create(@PathVariable Long questionId,@RequestBody ContentDto content, HttpSession session) {
-        return new AnswerDto(questionService.saveAnswer(session,content,questionId));
+    public AnswerDto create(@PathVariable Long questionId, @RequestBody Content content, HttpSession session) {
+        return new AnswerDto(questionService.saveAnswer(session, content, questionId));
     }
 
     @DeleteMapping("/{id}")
-    public AnswerDto delete(@PathVariable Long id ,HttpSession session) {
-        return new AnswerDto(questionService.deleteAnswer(id,session));
+    public AnswerDto delete(@PathVariable Long id, HttpSession session) {
+        return new AnswerDto(questionService.deleteAnswer(id, session));
     }
 
 }
